enh(Message): split MessageBody component out of Message
